# Performance Improvement: Eliminate calls to `toByteArray` when deserialise proto-buf