Protobuf can accept `Buffer` as input for `deserializeBinary`.

In a simple benchmark I did locally:
* ruby code, while true - send message 12kb size
* node code - subscribe to message (queue group, durable) 

The node process was at 100% CPU, after this change it’s reduced to 10% peak.

And bench-sub results:

| messages count | master - total time | master - msgs/sec | PR - total time | PR - msgs/sec |
| -------------- | ---------------------- | -------------------- | ----------------------- | --------------------- |
| 10,000 | 232ms | 43,103 | 163ms | 61,350 |
| 100,000 | 1,185ms | 84,388 | 790ms | 126,582 |
| 1,000,000 | 9,303ms | 107,492 | 5,999ms | 166,7222 |

all benchmarks running on:
node v10.15.3
MacBook Pro (15-inch, 2018)