@aricart 2019-07-08
I think I am going to release it for you right now. There's one more change that I need to do regarding the typescript definitions but that may take a bit of time.