@aricart 2019-07-08
@yosiat I released 0.2.4 for you